// contracts/typescript/models/RequestView.ts
export interface RequestView {
    /**
     *
     */
    acceptedProposalId: string | null;
    /**
     *
     */
    createdAt: Date;
    /**
     *
     */
    days: Array<RequestedDayView>;
    /**
     *
     */
    employeeId: string;
    /**
     *
     */
    id: string;
    /**
     *
     */
    note: string;
    /**
     *
     */
    parentRevisionId: string | null;
    /**
     *
     */
    revision: number;
    /**
     *
     */
    rootId: string;
    /**
     *
     */
    state: RequestState;
    /**
     *
     */
    submittedAt: Date | null;
    /**
     *
     */
    version: number;
}