// contracts/typescript/models/NotificationPage.ts
/* tslint:disable */
/* eslint-disable */
/**
 * HomeOfficeReservation API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { NotificationView } from './NotificationView';
import {
    NotificationViewFromJSON,
    NotificationViewFromJSONTyped,
    NotificationViewToJSON,
    NotificationViewToJSONTyped,
} from './NotificationView';

/**
 *
 * @export
 * @interface NotificationPage
 */
export interface NotificationPage {
    /**
     *
     */
    items: Array<NotificationView>;
    /**
     *
     */
    nextOffset: number | null;
    /**
     *
     */
    unreadCount: number;
}

/**
 * Check if a given object implements the NotificationPage interface.
 */
export function instanceOfNotificationPage(value: object): value is NotificationPage {
    if (!('items' in value) || value['items'] === undefined) return false;
    if (!('nextOffset' in value) || value['nextOffset'] === undefined) return false;
    if (!('unreadCount' in value) || value['unreadCount'] === undefined) return false;
    return true;
}

export function NotificationPageFromJSON(json: any): NotificationPage {
    return NotificationPageFromJSONTyped(json, false);
}

export function NotificationPageFromJSONTyped(json: any, ignoreDiscriminator: boolean): NotificationPage {
    if (json == null) {
        return json;
    }
    return {

        'items': ((json['items'] as Array<any>).map(NotificationViewFromJSON)),
        'nextOffset': json['nextOffset'],
        'unreadCount': json['unreadCount'],
    };
}

export function NotificationPageToJSON(json: any): NotificationPage {
    return NotificationPageToJSONTyped(json, false);
}

export function NotificationPageToJSONTyped(value?: NotificationPage | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {

        'items': ((value['items'] as Array<any>).map(NotificationViewToJSON)),
        'nextOffset': value['nextOffset'],
        'unreadCount': value['unreadCount'],
    };
}